Moved the WhoAmIProps in the models folder
