+++ src/components/routes/account.tsx
@@ -1,5 +1,5 @@
 import useAppUser from 'commons/components/app/hooks/useAppUser';
-import { CustomUser } from 'components/hooks/useMyUser';
+import { CustomUser } from 'components/models/ui/user';
 import User from 'components/routes/user';
 
 export default function Account() {
