+++ src/components/routes/admin/error_detail.tsx
@@ -11,8 +11,8 @@ import useAppUser from 'commons/components/app/hooks/useAppUser';
 import PageCenter from 'commons/components/pages/PageCenter';
 import useClipboard from 'commons/components/utils/hooks/useClipboard';
 import useMyAPI from 'components/hooks/useMyAPI';
-import { CustomUser } from 'components/hooks/useMyUser';
 import { Error } from 'components/models/base/error';
+import { CustomUser } from 'components/models/ui/user';
 import { DEFAULT_TAB, TAB_OPTIONS } from 'components/routes/file/viewer';
 import FileDownloader from 'components/visual/FileDownloader';
 import 'moment/locale/fr';
