+++ src/components/routes/admin/error_viewer.tsx
@@ -10,9 +10,9 @@ import PageHeader from 'commons/components/pages/PageHeader';
 import { useEffectOnce } from 'commons/components/utils/hooks/useEffectOnce';
 import useDrawer from 'components/hooks/useDrawer';
 import useMyAPI from 'components/hooks/useMyAPI';
-import { CustomUser } from 'components/hooks/useMyUser';
 import { Error } from 'components/models/base/error';
 import { SearchResult } from 'components/models/ui/search';
+import { CustomUser } from 'components/models/ui/user';
 import { ChipList } from 'components/visual/ChipList';
 import Histogram from 'components/visual/Histogram';
 import LineGraph from 'components/visual/LineGraph';
