+++ src/components/routes/admin/identify.tsx
@@ -5,7 +5,7 @@ import useAppUser from 'commons/components/app/hooks/useAppUser';
 import PageFullSize from 'commons/components/pages/PageFullSize';
 import { useEffectOnce } from 'commons/components/utils/hooks/useEffectOnce';
 import useMyAPI from 'components/hooks/useMyAPI';
-import { CustomUser } from 'components/hooks/useMyUser';
+import { CustomUser } from 'components/models/ui/user';
 import { RouterPrompt } from 'components/visual/RouterPrompt';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
