+++ src/components/routes/admin/service_review.tsx
@@ -17,9 +17,9 @@ import useAppUser from 'commons/components/app/hooks/useAppUser';
 import PageFullWidth from 'commons/components/pages/PageFullWidth';
 import { useEffectOnce } from 'commons/components/utils/hooks/useEffectOnce';
 import useMyAPI from 'components/hooks/useMyAPI';
-import { CustomUser } from 'components/hooks/useMyUser';
 import { Service as ServiceData } from 'components/models/base/service';
 import { ServiceStats as ServiceStatsData } from 'components/models/ui/service';
+import { CustomUser } from 'components/models/ui/user';
 import LineGraph from 'components/visual/LineGraph';
 import { getVersionQuery } from 'helpers/utils';
 import 'moment/locale/fr';
