+++ src/components/routes/admin/site_map.tsx
@@ -19,9 +19,9 @@ import useAppUser from 'commons/components/app/hooks/useAppUser';
 import PageFullWidth from 'commons/components/pages/PageFullWidth';
 import { useEffectOnce } from 'commons/components/utils/hooks/useEffectOnce';
 import useMyAPI from 'components/hooks/useMyAPI';
-import { CustomUser } from 'components/hooks/useMyUser';
 import { Role } from 'components/models/base/user';
 import { SiteMap as TSiteMap } from 'components/models/ui';
+import { CustomUser } from 'components/models/ui/user';
 import { PossibleColor } from 'components/models/utils/color';
 import CustomChip from 'components/visual/CustomChip';
 import { useState } from 'react';
