+++ src/components/routes/alerts/alert-details.tsx
@@ -23,8 +23,8 @@ import PageFullWidth from 'commons/components/pages/PageFullWidth';
 import useClipboard from 'commons/components/utils/hooks/useClipboard';
 import useALContext from 'components/hooks/useALContext';
 import useMyAPI from 'components/hooks/useMyAPI';
-import { CustomUser } from 'components/hooks/useMyUser';
 import { Alert as AlertData, DetailedItem } from 'components/models/base/alert';
+import { CustomUser } from 'components/models/ui/user';
 import { detailedItemCompare } from 'components/routes/alerts/hooks/useAlerts';
 import { ActionableChipList } from 'components/visual/ActionableChipList';
 import { ActionableCustomChipProps } from 'components/visual/ActionableCustomChip';
