+++ src/components/routes/authorize.tsx
@@ -4,8 +4,8 @@ import useAppBanner from 'commons/components/app/hooks/useAppBanner';
 import useAppUser from 'commons/components/app/hooks/useAppUser';
 import PageCardCentered from 'commons/components/pages/PageCardCentered';
 import useALContext from 'components/hooks/useALContext';
-import { CustomUser } from 'components/hooks/useMyUser';
 import { Role, Scope, SCOPES } from 'components/models/base/user';
+import { CustomUser } from 'components/models/ui/user';
 import getXSRFCookie from 'helpers/xsrf';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router';
