+++ src/components/routes/file/detail.tsx
@@ -1,6 +1,6 @@
 import useAppUser from 'commons/components/app/hooks/useAppUser';
 import PageCenter from 'commons/components/pages/PageCenter';
-import { CustomUser } from 'components/hooks/useMyUser';
+import { CustomUser } from 'components/models/ui/user';
 import FileDetail from 'components/visual/FileDetail';
 import { useParams } from 'react-router-dom';
 import ForbiddenPage from '../403';
