+++ src/components/routes/retrohunt/detail.tsx
@@ -22,10 +22,10 @@ import PageFullSize from 'commons/components/pages/PageFullSize';
 import useALContext from 'components/hooks/useALContext';
 import useDrawer from 'components/hooks/useDrawer';
 import useMyAPI from 'components/hooks/useMyAPI';
-import { CustomUser } from 'components/hooks/useMyUser';
 import { FileIndexed } from 'components/models/base/file';
 import { Retrohunt } from 'components/models/base/retrohunt';
 import { SearchResult } from 'components/models/ui/search';
+import { CustomUser } from 'components/models/ui/user';
 import ForbiddenPage from 'components/routes/403';
 import NotFoundPage from 'components/routes/404';
 import { RetrohuntPhase } from 'components/routes/retrohunt';
