+++ src/components/visual/Attack.tsx
@@ -6,8 +6,8 @@ import useAppUser from 'commons/components/app/hooks/useAppUser';
 import useClipboard from 'commons/components/utils/hooks/useClipboard';
 import useALContext from 'components/hooks/useALContext';
 import useHighlighter from 'components/hooks/useHighlighter';
-import { CustomUser } from 'components/hooks/useMyUser';
 import useSafeResults from 'components/hooks/useSafeResults';
+import { CustomUser } from 'components/models/ui/user';
 import CustomChip, { PossibleColors } from 'components/visual/CustomChip';
 import { safeFieldValueURI } from 'helpers/utils';
 import React, { useCallback } from 'react';
