+++ src/components/visual/FileDetail.tsx
@@ -25,9 +25,9 @@ import useAppUser from 'commons/components/app/hooks/useAppUser';
 import useALContext from 'components/hooks/useALContext';
 import useMyAPI from 'components/hooks/useMyAPI';
 import useMySnackbar from 'components/hooks/useMySnackbar';
-import { CustomUser } from 'components/hooks/useMyUser';
 import type { Error } from 'components/models/base/error';
 import type { File } from 'components/models/ui/file';
+import { CustomUser } from 'components/models/ui/user';
 import ForbiddenPage from 'components/routes/403';
 import { DEFAULT_TAB, TAB_OPTIONS } from 'components/routes/file/viewer';
 import AISummarySection from 'components/routes/submission/detail/ai_summary';
