+++ src/components/visual/FileViewer/hex.tsx
@@ -2,7 +2,7 @@ import { Alert, LinearProgress } from '@mui/material';
 import makeStyles from '@mui/styles/makeStyles';
 import useAppUser from 'commons/components/app/hooks/useAppUser';
 import useMyAPI from 'components/hooks/useMyAPI';
-import { CustomUser } from 'components/hooks/useMyUser';
+import { CustomUser } from 'components/models/ui/user';
 import ForbiddenPage from 'components/routes/403';
 import { HexViewerApp } from 'components/visual/HexViewer';
 import React, { useEffect, useState } from 'react';
